Create rbenv group before rbenv user
On first chef-client run during bootstrap, the user will be created,
and on many platforms will automagically have a group with the same
name created. The Ruby process chef-client is running under will not
be aware of this new group, so the call to create the group will fail.
Future runs will be fine.

By moving the group creation ahead of the user, we make sure Chef
knows about it and doesn't try to create it a second time.

See http://help.opscode.com/discussions/problems/742-group-error for
another approach, which we fortunately don't need here.