Adding in support for urls as patches.
Firstly this requires patchutils to be installed so that we can use
filterdiff. I’m not entirely happy with this, but is a first cut of
this functionality so happy to have a discussion over it.

Allow the rbenv_ruby LWRP to support a patch variable, which is simply
a URL that rbenv install --patch will pull in. Written in a similar
style to here.
https://github.com/sstephenson/ruby-build/wiki#make-error-for-200-p247-a
nd-lower-on-fedorared-hat