Ensure $HOME is correct before a ruby build (possibly from git.)
This is a continuation of the work in db913928da8a389ff8edd98309dfb00ea3fa9fd8 which added a workaround for a git bug when using git to clone the rbenv install. But the bug is also present when a ruby build happens from a git source. This patch ensures that $HOME is temporarily changed to the home directory for `node[:rbenv][:user]` during the ruby build process.