Unset TMPDIR on Darwin (Mac OS X)
Between Chef, Ruby, and the rbenv cookbook, there is a edge case that
manifests on OS X when you set an rbenv user.

It is caused by the fact that the rbenv command is run under the context
of the rbenv user, but still has some of the shell environments from the
initiating user (like TMPDIR), which means that rbenv can't write to the
temporary directory, so we can't let the rbenv commands rely on that
environment variable.

This simply ensures that we use a standard temp directory and adds
"mac_os_x" to the supported operating systems.