Check for successful ruby install under rbenv.
If the URL provided for the install is not available (for example, flaky internet, behind a corporate firewall), the first invocation of this cookbook creates an rbenv directory, and then fails to :sync. Subsequent invocations of the cookbook see that extant directory, and assume that rbenv has been installed, without needing to :sync, and fail. This edit looks for at least one rbenv/versions/*/bin/ruby file.